+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Allow invoices to be marked as sent
 - Support for Profit & Loss Report 
 - Support for Tax Summary Report
+- Support for Payments Collected Report
 
 ### 3.0.0